pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "animation_in_terminal"
version = "0.1.0"
description = "Terminal-based animation display"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "textual~=0.85.1",
    "pyfiglet~=1.0.2",
    "rich~=13.9.3",
    "rich_pixels~=3.0.1",
    "pygame-ce~=2.5.1",
]

[project.scripts]
animate = "animation.app:main"

[tool.setuptools.package-data]
animation = [
    "styles/*.tcss",
    "assets/sounds/*.wav",
    "assets/frames/*.png"
]

[tool.setuptools.packages.find]
where = ["src"]
